CAMEL-13374: Move XmlStreamReader into camel-xml-io-util module to be independent.
