+++ core/camel-xml-io-util/src/main/java/org/apache/camel/xml/io/util/XmlStreamReaderException.java
@@ -33,7 +33,7 @@
  */
 
 // CHECKSTYLE:OFF
-package org.apache.camel.xml.io;
+package org.apache.camel.xml.io.util;
 
 import java.io.IOException;
 import java.io.InputStream;
