+++ core/camel-xml-io/pom.xml
@@ -27,7 +27,6 @@
     </parent>
 
     <artifactId>camel-xml-io</artifactId>
-
     <name>Camel :: XML IO</name>
     <description>Camel XML IO</description>
 
@@ -40,6 +39,10 @@
         <dependency>
             <groupId>org.apache.camel</groupId>
             <artifactId>camel-core-engine</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel</groupId>
+            <artifactId>camel-xml-io-util</artifactId>
         </dependency>
 
         <!-- testing -->
