+++ core/camel-xml-io/src/main/java/org/apache/camel/xml/io/MXParser.java
@@ -33,6 +33,8 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 
+import org.apache.camel.xml.io.util.XmlStreamReader;
+
 //TODO best handling of interning issues
 //   have isAllNewStringInterned ???
 
@@ -46,7 +48,6 @@ import java.io.Reader;
  * @author <a href="http://www.extreme.indiana.edu/~aslom/">Aleksander
  *         Slominski</a>
  */
-
 public class MXParser implements XmlPullParser {
     // NOTE: no interning of those strings --> by Java lang spec they MUST be
     // already interned
